firmware: move ice40 driver instance to ice40_spi.x
