+++ firmware/common/fpga_image.c
@@ -23,7 +23,6 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include "drivers.h"
 #include "fpga.h"
 #include "ice40_spi.h"
 #include "lz4_blk.h"
