+++ firmware/common/ice40_spi.c
@@ -28,6 +28,23 @@
 #include "delay.h"
 #include "platform_scu.h"
 
+/* Driver instance. */
+ssp_config_t ssp_config_ice40_fpga = {
+	.data_bits = SSP_DATA_8BITS,
+	.spi_mode = SSP_CPOL_1_CPHA_1,
+	.serial_clock_rate = 21,
+	.clock_prescale_rate = 2,
+};
+
+ice40_spi_driver_t ice40 = {
+	.bus = &spi_bus_ssp1,
+};
+
+void ssp1_set_mode_ice40(void)
+{
+	spi_bus_start(&spi_bus_ssp1, &ssp_config_ice40_fpga);
+}
+
 void ice40_spi_target_init(ice40_spi_driver_t* const drv)
 {
 	const platform_scu_t* scu = platform_scu();
